+++ src/blocks/container/edit.js
@@ -1,52 +1,31 @@
 import BlockControls from './components/BlockControls';
 import InspectorAdvancedControls from '../grid/components/InspectorAdvancedControls';
-import ComponentCSS from './components/ComponentCSS';
 import GoogleFontLink from '../../components/google-font-link';
-import Element from '../../components/element';
-import RootElement from '../../components/root-element';
 import { applyFilters } from '@wordpress/hooks';
-import { InnerBlocks, useBlockProps } from '@wordpress/block-editor';
-import { Button } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-import getIcon from '../../utils/get-icon';
 import { Fragment, useEffect } from '@wordpress/element';
-import { useDeviceType, useInnerBlocksCount } from '../../hooks';
-import classnames from 'classnames';
-import ShapeDividers from './components/ShapeDividers';
+import { useDeviceType } from '../../hooks';
 import InspectorControls from './components/InspectorControls';
-import GridItem from './components/GridItem';
 import { compose } from '@wordpress/compose';
 import { withUniqueId, withContainerLegacyMigration } from '../../hoc';
-import { useDispatch } from '@wordpress/data';
 import withDynamicContent from '../../extend/dynamic-content/hoc/withDynamicContent';
+import ContainerContentRenderer from './components/ContainerContentRenderer';
 
 const ContainerEdit = ( props ) => {
 	const {
 		attributes,
 		setAttributes,
 		clientId,
-		name,
+		ContentRenderer = ContainerContentRenderer,
 	} = props;
 
 	const {
-		uniqueId,
-		className,
 		anchor,
-		tagName,
-		backgroundColor,
 		fontFamily,
 		googleFont,
 		googleFontVariants,
-		isGrid,
-		bgOptions,
-		bgImage,
-		bgImageInline,
 	} = attributes;
 
-	const { selectBlock } = useDispatch( 'core/block-editor' );
 	const [ deviceType, setDeviceType ] = useDeviceType( 'Desktop' );
-	const innerBlocksCount = useInnerBlocksCount( clientId );
-	const hasChildBlocks = 0 < innerBlocksCount;
 
 	useEffect( () => {
 		const thisBlock = document.getElementById( `block-${ clientId }` );
@@ -82,52 +61,6 @@ const ContainerEdit = ( props ) => {
 	);
 
 	const filterTagName = ( tagValue ) => allowedTagNames.includes( tagValue ) ? tagValue : 'div';
-
-	let hasStyling = (
-		!! backgroundColor ||
-		attributes.borderSizeTop ||
-		attributes.borderSizeRight ||
-		attributes.borderSizeBottom ||
-		attributes.borderSizeLeft
-	);
-
-	hasStyling = applyFilters( 'generateblocks.editor.containerHasStyling', hasStyling, props );
-
-	let htmlAttributes = {
-		className: classnames( {
-			'gb-container': true,
-			[ `gb-container-${ uniqueId }` ]: true,
-			[ `${ className }` ]: undefined !== className,
-			'gb-container-empty': ! hasChildBlocks,
-			'gb-container-visual-guides': ! hasChildBlocks && ! hasStyling && ! props.isSelected,
-		} ),
-		id: anchor ? anchor : null,
-	};
-
-	if ( bgImageInline && bgImage.image && bgImage.image.url ) {
-		const backgroundUrl = applyFilters( 'generateblocks.editor.bgImageURL', bgImage.image.url, props );
-
-		if ( backgroundUrl ) {
-			let imageAttributeName = 'background-image';
-
-			if ( 'element' !== bgOptions.selector ) {
-				imageAttributeName = '--' + imageAttributeName;
-			}
-
-			htmlAttributes.style = {
-				[ imageAttributeName ]: 'url(' + backgroundUrl + ')',
-			};
-		}
-	}
-
-	htmlAttributes = applyFilters(
-		'generateblocks.frontend.htmlAttributes',
-		htmlAttributes,
-		'generateblocks/container',
-		attributes
-	);
-
-	const blockProps = useBlockProps( htmlAttributes );
 
 	const allShapes = [];
 
@@ -184,55 +117,19 @@ const ContainerEdit = ( props ) => {
 
 			<InspectorAdvancedControls anchor={ anchor } setAttributes={ setAttributes } />
 
-			<ComponentCSS { ...props } deviceType={ deviceType } />
-
 			<GoogleFontLink
 				fontFamily={ fontFamily }
 				googleFont={ googleFont }
 				googleFontVariants={ googleFontVariants }
 			/>
 
-			<RootElement name={ name } clientId={ clientId }>
-				<GridItem isGrid={ isGrid } uniqueId={ uniqueId }>
-					<Element
-						tagName={ filterTagName( applyFilters( 'generateblocks.frontend.containerTagName', tagName, attributes ) ) }
-						htmlAttrs={ blockProps }
-					>
-						{ applyFilters( 'generateblocks.frontend.afterContainerOpen', '', attributes ) }
-						<div className={ 'gb-inside-container' }>
-							{ applyFilters( 'generateblocks.frontend.insideContainer', '', attributes ) }
-							<InnerBlocks
-								templateLock={ false }
-								renderAppender={ () => {
-									// Selected Container.
-									if ( props.isSelected ) {
-										return <InnerBlocks.ButtonBlockAppender />;
-									}
-
-									// Empty non-selected Container.
-									if ( ! hasChildBlocks && ! props.isSelected ) {
-										return <Button
-											className="gblocks-container-selector"
-											onClick={ () => selectBlock( clientId ) }
-											aria-label={ __( 'Select Container', 'generateblocks' ) }
-										>
-											<span className="gblocks-container-selector__icon">
-												{ getIcon( 'container' ) }
-											</span>
-										</Button>;
-									}
-
-									return false;
-								} }
-							/>
-						</div>
-
-						<ShapeDividers attributes={ attributes } allShapes={ allShapes } />
-
-						{ applyFilters( 'generateblocks.frontend.beforeContainerClose', '', attributes ) }
-					</Element>
-				</GridItem>
-			</RootElement>
+			<ContentRenderer
+				{ ...props }
+				generateBlocksInfo={ generateBlocksInfo }
+				filterTagName={ filterTagName }
+				allShapes={ allShapes }
+				deviceType={ deviceType }
+			/>
 		</Fragment>
 	);
 };
